SAMD21 Cortex M0+ Support
Testing thison Seeeduino Xiao and in my experience it's digital pins must be declared as a static const int or weird issues arise.

Additionally, the SAMD21 compiler will error out due to line 13... It will not accept an int of 0x0, seems to consider that a null value. Adding (byte) operator allows the code to compile.

The rest of the changes are semantic just so I can visualize what's going on in my head. Not meant for main repo.

Not a fan of default combinations in any context; I changed the numbers to the Master Lock/Sentry HQ ZIP code 53210.